added code for displaying output aggregation, added new rules
after testing with the exist member ship functions the triangular relationship did not work for the extreme values cases (i.e. age is 130, headache is 10, temperature is 45) and would cause NaN values to be output by the centroid function. After changing the low and high cases for the decision variables to trapezoidal functions this unwanted behavior ceased. More testing is required before stating as complete, but this is a start